split obj into multiple files
