+++ constants.py
@@ -1,0 +1,10 @@
+SCREENH = 600
+SCREENW = 640
+LEFT = 0
+RIGHT = 1
+UP = 2
+DOWN = 3
+BLACK = (0, 0, 0)
+EXTIMELAPSE = 250
+BOSSHEALTH = 2000  # default 2000
+PLAYERHEALTH = 3  # default 3
