+++ loadstaticres.py
@@ -1,0 +1,5 @@
+from utilfuncs import toframes
+import pygame
+
+
+explosion = toframes(pygame.image.load("explode1.png"), 5, 120)
